Moved modals back to App. It won't display properly if it's in Footer.
